fix cursor issue
