+++ packages/workspace-store/src/persistence/index.ts
@@ -36,16 +36,6 @@ type WorkspaceStoreShape = {
 }
 
 /**
- * Composes the legacy `${teamSlug}/${slug}` workspace id. Kept for callers
- * that still build router-level identifiers from slug pairs (for example
- * synthetic picker placeholders for not-yet-persisted team workspaces).
- *
- * The persistence layer itself no longer uses this format — chunk records
- * are keyed by `workspaceUid`. Prefer `workspaceUid` whenever you have it.
- */
-export const getWorkspaceId = (teamSlug: string, slug: string) => `${teamSlug}/${slug}`
-
-/**
  * Generates a fresh `workspaceUid` for new workspaces.
  *
  * Wraps `crypto.randomUUID` so every caller produces UIDs in a consistent
